js refactoring
