+++ src/js/components/NewsCard.js
