+++ src/js/components/SearchStatus.js
